solving issue where scale wouldn't be taken into account on sprite

+++ PeanutButter/src/PeanutButter/Entity/Component/Sprite.h
@@ -80,8 +80,8 @@ namespace PeanutButter {
 
 			DestinationRectangle.x = static_cast<int>(transform->Position->x) - Application::s_Camera.x;
 			DestinationRectangle.y = static_cast<int>(transform->Position->y) - Application::s_Camera.y;
-			DestinationRectangle.w = static_cast<int>(SpriteSize.x) * static_cast<int>(transform->Scale->x);
-			DestinationRectangle.h = static_cast<int>(SpriteSize.y) * static_cast<int>(transform->Scale->y);
+			DestinationRectangle.w = static_cast<int>(SpriteSize.x) * transform->Scale->x;
+			DestinationRectangle.h = static_cast<int>(SpriteSize.y) * transform->Scale->y;
 		}
 
 		void Render() override {
